requirements.txt
openai
pyyaml
faiss-cpu
sentence-transformers
streamlit
matplotlib
chromadb
tqdm
flask
rich
textual
pandas
requests
beautifulsoup4
python-dotenv
astor
pyttsx3

# 🚀 Production & Advanced AGI Modules
gunicorn
uvicorn[standard]
langchain
